pyproject.toml
[tool.poetry]
name = "StockMeme"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.11"
yfinance = "^0.2.37"
matplotlib = "^3.8.3"
pandas = "^2.2.1"
discord-py = "^2.3.2"

[tool.poetry.dev-dependencies]
black = "24.2.0"
ptpython = ">= 3.0.26"

[tool.black]
line-length = 150
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["ALL"]
ignore = [
    "COM812",
    "D200",
    "D212",
    "D400",
    "D403",
    "D415",
    "ERA001",
    "I001",
    "S104",
    "T201",
]
target-version = "py311"
